+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - add berkeley-qa support (mina-signer-experimental)
 - update auro-web3-provider to @aurowallet/mina-provider
 - add berkeley-qa signParty
+- add gql params type
 ## [2.0.4]
 
 - Re-encode raw signature to fix ledger transation signature  (https://github.com/jspada/ledger-app-mina/pull/22)