+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - update auro-web3-provider to @aurowallet/mina-provider
 - add berkeley-qa signParty
 - add gql params type
+- set default net type to Berkeley-QA
 ## [2.0.4]
 
 - Re-encode raw signature to fix ledger transation signature  (https://github.com/jspada/ledger-app-mina/pull/22)